refactor: extract FunctionTypeValidation mixin to remove SUPPORTED_TYPES duplication
ScalarFunction and AggregateFunction shared an identical SUPPORTED_TYPES
constant and a private check_supported_type! helper. Extract both into a
new DuckDB::FunctionTypeValidation module and include it in both classes.

- Add lib/duckdb/function_type_validation.rb with SUPPORTED_TYPES and
  check_supported_type! (private)
- Remove the duplicated constant and method from scalar_function.rb and
  aggregate_function.rb
- Require the new module in duckdb.rb before the two function files
- All 128 ScalarFunction + 14 AggregateFunction tests still pass
- rubocop: no offenses
