+++ lib/duckdb.rb
@@ -14,6 +14,7 @@ require 'duckdb/appender'
 require 'duckdb/config'
 require 'duckdb/column'
 require 'duckdb/logical_type'
+require 'duckdb/function_type_validation'
 require 'duckdb/scalar_function'
 require 'duckdb/scalar_function_set'
 require 'duckdb/aggregate_function'
